refactor: move write/read package json into JsPackageManager
