+++ code/lib/cli/src/js-package-manager/Yarn2Proxy.test.ts
@@ -1,5 +1,4 @@
 import { Yarn2Proxy } from './Yarn2Proxy';
-import * as PackageJsonHelper from './PackageJsonHelper';
 
 describe('Yarn 2 Proxy', () => {
   let yarn2Proxy: Yarn2Proxy;
@@ -62,7 +61,7 @@ describe('Yarn 2 Proxy', () => {
     it('skipInstall should only change package.json without running install', () => {
       const executeCommandSpy = jest.spyOn(yarn2Proxy, 'executeCommand').mockReturnValue('7.0.0');
       const writePackageSpy = jest
-        .spyOn(PackageJsonHelper, 'writePackageJson')
+        .spyOn(yarn2Proxy, 'writePackageJson')
         .mockImplementation(jest.fn);
 
       yarn2Proxy.removeDependencies(
