+++ code/lib/cli/src/js-package-manager/index.ts
@@ -1,4 +1,3 @@
 export * from './JsPackageManagerFactory';
 export * from './JsPackageManager';
 export * from './PackageJson';
-export * from './PackageJsonHelper';
